requirements.txt
gunicorn
flask
simplejson
numpy
scipy
keras
h5py
pillow
pyyaml
tensorflow
